+++ Runtime/Scripts/Api/NmkrApiTools.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading.Tasks;
 using Nmkr.Sdk.Schemas;
 
 namespace Nmkr.Sdk
@@ -135,10 +136,10 @@ namespace Nmkr.Sdk
         /// <param name="address">The wallet address to check for assets</param>
         /// <param name="onSuccess"><include file='SummaryParams.xml' path='docs/param[@name="onSuccess"]/*' /></param>
         /// <param name="onFailure"><include file='SummaryParams.xml' path='docs/param[@name="onFailure"]/*' /></param>
-        public static async void GetAllAssetsInWallet(string address, Action<AssetsAssociatedWithAccount[]> onSuccess, Action<ResponseError> onFailure = null)
+        public static async Task<ApiResponse<AssetsAssociatedWithAccount[]>> GetAllAssetsInWallet(string address, Action<AssetsAssociatedWithAccount[]> onSuccess = null, Action<ResponseError> onFailure = null)
         {
             var endpoint = $"GetAllAssetsInWallet/{address}";
-            await GetAsync(endpoint, onSuccess, onFailure);
+            return await GetAsync(endpoint, onSuccess, onFailure);
         }
     }
 }